Move ChartParameters out of shared/*
